exception handling done

+++ src/main/java/com/assignment/slisAPI/service/StudentService.java
@@ -2,6 +2,7 @@ package com.assignment.slisAPI.service;
 
 import java.util.List;
 
+import com.assignment.slisAPI.exception.ResourceNotFoundException;
 import com.assignment.slisAPI.model.Student;
 import com.assignment.slisAPI.repository.StudentRepository;
 
@@ -17,27 +18,45 @@ public class StudentService {
         return repository.save(student);
     }
 
-    public Student getStudentById(int id) {
-        return repository.findById(id).orElse(null);
+    public Student getStudentById(int id) throws ResourceNotFoundException {
+        Student student = repository.findById(id).orElse(null);
+        if (student != null) {
+            return student;
+        } else {
+            throw new ResourceNotFoundException("Student not found with Id:" + id);
+        }
     }
-
-    public Student getStudentByName(String name) {
-        return repository.findByName(name);
+    
+    public Student getStudentByName(String name) throws ResourceNotFoundException {
+        Student student = repository.findByName(name);
+        if (student != null) {
+            return student;
+        } else {
+            throw new ResourceNotFoundException("Student not found with name :" + name);
+        }
     }
 
     public List<Student> getAllStudent() {
         return repository.findAll();
     }
 
-    public String deleteStudentById(int id) {
-        repository.deleteById(id);
-        return "student removed " + id;
+    public String deleteStudentById(int id)  throws ResourceNotFoundException {
+        if (repository.findById(id).orElse(null) != null) {
+            repository.deleteById(id);
+            return "student removed " + id;
+        } else {
+            throw new ResourceNotFoundException("Student not found with Id :" + id);
+        }
     }
 
-    public Student updateStudent(Student student) {
+    public Student updateStudent(Student student) throws ResourceNotFoundException {
         Student exStudent = repository.findById(student.getId()).orElse(null);
-        exStudent.setName(student.getName());
-        return repository.save(exStudent);
+        if (exStudent != null) {
+            exStudent.setName(student.getName());
+            return repository.save(exStudent);
+        } else {
+            throw new ResourceNotFoundException("Student not found with Id :" + student.getId());
+        }
     }
 
 }
